+++ src/connectors/scanner/mod.rs
@@ -1,4 +1,4 @@
-use crate::connectors::metadata::SourceMetadata;
+use crate::connectors::metadata::FileLikeMetadata;
 use crate::connectors::ReadError;
 use crate::persistence::cached_object_storage::CachedObjectStorage;
 
@@ -13,8 +13,8 @@ pub use s3::S3Scanner;
 
 #[derive(Clone, Debug)]
 pub enum QueuedAction {
-    Read(Vec<u8>, SourceMetadata),
-    Update(Vec<u8>, SourceMetadata),
+    Read(Vec<u8>, FileLikeMetadata),
+    Update(Vec<u8>, FileLikeMetadata),
     Delete(Vec<u8>),
 }
 
@@ -28,7 +28,10 @@ impl QueuedAction {
 
 #[allow(clippy::module_name_repetitions)]
 pub trait PosixLikeScanner: Send {
-    fn object_metadata(&mut self, object_path: &[u8]) -> Result<Option<SourceMetadata>, ReadError>;
+    fn object_metadata(
+        &mut self,
+        object_path: &[u8],
+    ) -> Result<Option<FileLikeMetadata>, ReadError>;
     fn read_object(&mut self, object_path: &[u8]) -> Result<Vec<u8>, ReadError>;
     fn next_scanner_actions(
         &mut self,